Set channel_count and sample_rate in stream_configuration.rs
Problem: I'm not able to set `channel_count` and `sample_rate` in StreamConfiguration as a consumer of this library.

Solution: Write up the setter calls to the actual operations to set the aforementioned attributes via objc interface.

Testing: Tested on my Macbook. I'm able to update the aforementioned attributes and see the changes reflected in the channel count and sample rate of the audio samples that I'm receiving in `did_output_sample_buffer` callback.

Signed-off-by: Alex Jiao <uohzxela@gmail.com>